requirements.txt
fastapi
uvicorn[standard]
jinja2
openai-whisper
soundcard
soundfile
numpy<2
